fixed comments code
