+++ src/validations/commentValidation.ts
@@ -2,7 +2,6 @@ import Joi from "joi";
 import { bookIdSchema } from "./bookValidation";
 
 export const commentJoiSchema = Joi.object({
-  user: Joi.string(),
   book: Joi.string().required(),
   text: Joi.string().required(),
 });
